# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Camila SaaS - API para emprendedores locales
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ErrorResponseDetails = { [key: string]: unknown };

export interface ErrorResponse {
  error: string;
  details?: ErrorResponseDetails;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  superadmin: "superadmin",
  store_admin: "store_admin",
  store_staff: "store_staff",
  cashier: "cashier",
} as const;

export interface User {
  id: string;
  storeId: string;
  name: string;
  email: string;
  role: UserRole;
  isActive: boolean;
  createdAt: string;
}

export type StoreBusinessType =
  (typeof StoreBusinessType)[keyof typeof StoreBusinessType];

export const StoreBusinessType = {
  clothing: "clothing",
  restaurant: "restaurant",
  bakery: "bakery",
  fair_booth: "fair_booth",
  general_catalog: "general_catalog",
} as const;

export type StoreDocumentType =
  (typeof StoreDocumentType)[keyof typeof StoreDocumentType];

export const StoreDocumentType = {
  DNI: "DNI",
  RUC10: "RUC10",
  RUC20: "RUC20",
} as const;

export type LicenseStatus = (typeof LicenseStatus)[keyof typeof LicenseStatus];

export const LicenseStatus = {
  trial: "trial",
  active: "active",
  expired: "expired",
  suspended: "suspended",
} as const;

export interface License {
  id: string;
  storeId: string;
  status: LicenseStatus;
  startsAt?: string;
  expiresAt?: string;
  notes?: string;
  updatedAt: string;
}

export interface Store {
  id: string;
  businessName: string;
  businessType: StoreBusinessType;
  documentType: StoreDocumentType;
  documentNumber: string;
  ownerName: string;
  phone: string;
  email: string;
  address?: string;
  district: string;
  logoUrl?: string;
  bannerUrl?: string;
  primaryColor?: string;
  description?: string;
  whatsapp?: string;
  socialInstagram?: string;
  socialFacebook?: string;
  isActive: boolean;
  license?: License;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  store: Store;
  token: string;
}

export interface CurrentUserResponse {
  user: User;
  store: Store;
  license?: License;
}

export type RegisterStoreRequestBusinessType =
  (typeof RegisterStoreRequestBusinessType)[keyof typeof RegisterStoreRequestBusinessType];

export const RegisterStoreRequestBusinessType = {
  clothing: "clothing",
  restaurant: "restaurant",
  bakery: "bakery",
  fair_booth: "fair_booth",
  general_catalog: "general_catalog",
} as const;

export type RegisterStoreRequestDocumentType =
  (typeof RegisterStoreRequestDocumentType)[keyof typeof RegisterStoreRequestDocumentType];

export const RegisterStoreRequestDocumentType = {
  DNI: "DNI",
  RUC10: "RUC10",
  RUC20: "RUC20",
} as const;

export interface RegisterStoreRequest {
  businessName: string;
  businessType: RegisterStoreRequestBusinessType;
  documentType: RegisterStoreRequestDocumentType;
  documentNumber: string;
  ownerName: string;
  phone: string;
  email: string;
  /** @minLength 8 */
  password: string;
  address?: string;
  district: string;
  logoUrl?: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface ForgotPasswordRequest {
  email: string;
}

export interface ResetPasswordRequest {
  token: string;
  /** @minLength 8 */
  password: string;
}

export interface UpdateStoreRequest {
  businessName?: string;
  address?: string;
  district?: string;
  phone?: string;
  logoUrl?: string;
  bannerUrl?: string;
  primaryColor?: string;
  description?: string;
  whatsapp?: string;
  socialInstagram?: string;
  socialFacebook?: string;
}

export interface AdminUpdateStoreRequest {
  businessName?: string;
  address?: string;
  district?: string;
  phone?: string;
  isActive?: boolean;
}

export type AddUserRequestRole =
  (typeof AddUserRequestRole)[keyof typeof AddUserRequestRole];

export const AddUserRequestRole = {
  store_admin: "store_admin",
  store_staff: "store_staff",
  cashier: "cashier",
} as const;

export interface AddUserRequest {
  name: string;
  email: string;
  /** @minLength 8 */
  password: string;
  role: AddUserRequestRole;
}

export type UpdateLicenseRequestStatus =
  (typeof UpdateLicenseRequestStatus)[keyof typeof UpdateLicenseRequestStatus];

export const UpdateLicenseRequestStatus = {
  trial: "trial",
  active: "active",
  expired: "expired",
  suspended: "suspended",
} as const;

export interface UpdateLicenseRequest {
  status: UpdateLicenseRequestStatus;
  startsAt?: string;
  expiresAt?: string;
  notes?: string;
}

export interface PaginatedStores {
  data: Store[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface AdminStats {
  totalStores: number;
  activeStores: number;
  trialStores: number;
  expiredStores: number;
  suspendedStores: number;
  newThisMonth: number;
}

export interface Category {
  id: string;
  storeId: string;
  name: string;
  description?: string;
  sortOrder?: number;
  createdAt: string;
}

export interface CreateCategoryRequest {
  name: string;
  description?: string;
  sortOrder?: number;
}

export interface Product {
  id: string;
  storeId: string;
  categoryId?: string;
  name: string;
  description?: string;
  price: number;
  costPrice?: number;
  sku?: string;
  barcode?: string;
  imageUrl?: string;
  stock: number;
  minStock?: number;
  unit?: string;
  isActive: boolean;
  createdAt: string;
  category?: Category;
}

export interface CreateProductRequest {
  name: string;
  description?: string;
  price: number;
  costPrice?: number;
  categoryId?: string;
  sku?: string;
  barcode?: string;
  imageUrl?: string;
  stock?: number;
  minStock?: number;
  unit?: string;
  isActive?: boolean;
}

export interface PaginatedProducts {
  data: Product[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export type AdjustInventoryRequestType =
  (typeof AdjustInventoryRequestType)[keyof typeof AdjustInventoryRequestType];

export const AdjustInventoryRequestType = {
  in: "in",
  out: "out",
  adjustment: "adjustment",
} as const;

export interface AdjustInventoryRequest {
  productId: string;
  quantity: number;
  type: AdjustInventoryRequestType;
  reason?: string;
  notes?: string;
}

export type InventoryMovementType =
  (typeof InventoryMovementType)[keyof typeof InventoryMovementType];

export const InventoryMovementType = {
  in: "in",
  out: "out",
  adjustment: "adjustment",
} as const;

export interface InventoryMovement {
  id: string;
  storeId: string;
  productId: string;
  userId: string;
  type: InventoryMovementType;
  quantity: number;
  previousStock: number;
  newStock: number;
  reason?: string;
  notes?: string;
  createdAt: string;
}

export interface PaginatedInventoryMovements {
  data: InventoryMovement[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export type AdminGetAllStoresParams = {
  page?: number;
  limit?: number;
  status?: AdminGetAllStoresStatus;
  search?: string;
};

export type AdminGetAllStoresStatus =
  (typeof AdminGetAllStoresStatus)[keyof typeof AdminGetAllStoresStatus];

export const AdminGetAllStoresStatus = {
  trial: "trial",
  active: "active",
  expired: "expired",
  suspended: "suspended",
} as const;

export type GetProductsParams = {
  page?: number;
  limit?: number;
  categoryId?: string;
  search?: string;
  lowStock?: boolean;
};

export type GetInventoryMovementsParams = {
  productId?: string;
  page?: number;
};