# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Camila SaaS - API para emprendedores locales
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categories:
    get:
      operationId: getCategories
      tags: [categories]
      summary: Get product categories for current store
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a new category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryRequest"
      responses:
        "201":
          description: Category created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{categoryId}:
    patch:
      operationId: updateCategory
      tags: [categories]
      summary: Update a category
      parameters:
        - name: categoryId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryRequest"
      responses:
        "200":
          description: Category updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete a category
      parameters:
        - name: categoryId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Category deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /products:
    get:
      operationId: getProducts
      tags: [products]
      summary: Get products for current store
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: categoryId
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: lowStock
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Paginated products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedProducts"

    post:
      operationId: createProduct
      tags: [products]
      summary: Create a new product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductRequest"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{productId}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a single product
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Product details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductRequest"
      responses:
        "200":
          description: Product updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Product deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /inventory/adjust:
    post:
      operationId: adjustInventory
      tags: [inventory]
      summary: Adjust inventory for a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdjustInventoryRequest"
      responses:
        "200":
          description: Inventory adjusted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryMovement"
  /inventory/movements:
    get:
      operationId: getInventoryMovements
      tags: [inventory]
      summary: Get inventory movement history
      parameters:
        - name: productId
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: Inventory movements
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedInventoryMovements"
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    Category:
      type: object
      properties:
        id:
          type: string
        storeId:
          type: string
        name:
          type: string
        description:
          type: string
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - storeId
        - name
        - createdAt
    CreateCategoryRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        sortOrder:
          type: integer
      required:
        - name
    Product:
      type: object
      properties:
        id:
          type: string
        storeId:
          type: string
        categoryId:
          type: string
        name:
          type: string
        description:
          type: string
        price:
          type: number
        costPrice:
          type: number
        sku:
          type: string
        barcode:
          type: string
        imageUrl:
          type: string
        stock:
          type: integer
        minStock:
          type: integer
        unit:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        category:
          $ref: "#/components/schemas/Category"
      required:
        - id
        - storeId
        - name
        - price
        - stock
        - isActive
        - createdAt
    CreateProductRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        costPrice:
          type: number
        categoryId:
          type: string
        sku:
          type: string
        barcode:
          type: string
        imageUrl:
          type: string
        stock:
          type: integer
          default: 0
        minStock:
          type: integer
          default: 0
        unit:
          type: string
        isActive:
          type: boolean
          default: true
      required:
        - name
        - price
    PaginatedProducts:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - data
        - total
        - page
        - limit
        - totalPages
    AdjustInventoryRequest:
      type: object
      properties:
        productId:
          type: string
        quantity:
          type: integer
        type:
          type: string
          enum: [in, out, adjustment]
        reason:
          type: string
        notes:
          type: string
      required:
        - productId
        - quantity
        - type
    InventoryMovement:
      type: object
      properties:
        id:
          type: string
        storeId:
          type: string
        productId:
          type: string
        userId:
          type: string
        type:
          type: string
          enum: [in, out, adjustment]
        quantity:
          type: integer
        previousStock:
          type: integer
        newStock:
          type: integer
        reason:
          type: string
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - storeId
        - productId
        - userId
        - type
        - quantity
        - previousStock
        - newStock
        - createdAt
    PaginatedInventoryMovements:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/InventoryMovement"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - data
        - total
        - page
        - limit
        - totalPages